rename jwt files to verification replace get method to put method and working verification
