+++ packages/py-api/py_api/routes/verification_routes.py
@@ -6,6 +6,6 @@ from py_api.controllers import VerificationController as c
 router = APIRouter(prefix="/hackathon/verify")
 
 
-@router.get("", response_model=Tuple[Dict[str, str], int])
-def verify_user_route(verification_token: str) -> Tuple[Dict[str, str], int]:
+@router.put("", response_model=Tuple[Dict[str, str], int])
+def verify_participant_route(verification_token: str) -> Tuple[Dict[str, str], int]:
     return c.verify_participant(verification_token)
